// src/packages/types/carservice/Ser_Inv_StockIn.ts
export interface Ser_Inv_StockInDtl {
  StockInDetailIndex: string;
  Quantity: string;
  Description: string;
  StockOutOrderNo: string;
  LogLUDateTime: string;
  LogLUBy: string;
  PartID: string;
  PartCode: string;
  VieName: string;
  Unit: string;
  StockInID: string;
  StockInNO: string;
  DealerCode: string;
  PlanLocationID: string;
  ActualLocationID: string;
  Price: string;
  VAT: string;
  InStock: string;
  BeforeTax: string;
  AfterTax: string;
}